# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ConeShare API
  version: 1.6.0
  description: OpenAPI schema for the ConeShare backend API.
paths:
...
  /api/v1/datarooms/:
    get:
      operationId: v1_datarooms_list
      tags:
      - datarooms
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dataroom'
          description: ''
    post:
      operationId: v1_datarooms_create
      tags:
      - datarooms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dataroom'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Dataroom'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Dataroom'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataroom'
          description: ''
...
  /api/v1/share-links/:
    get:
      operationId: v1_share_links_list
      tags:
      - sharelinks
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareLink'
          description: ''
    post:
      operationId: v1_share_links_create
      tags:
      - sharelinks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareLink'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ShareLink'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ShareLink'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'
          description: ''
...
components:
  schemas:
...
    Dataroom:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 255
        organization:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        created_by:
          type: string
          readOnly: true
          nullable: true
        show_file_index:
          type: boolean
        branding_banner:
          type: string
          format: uri
          nullable: true
        brand_primary_color:
          type: string
          nullable: true
          maxLength: 9
        brand_secondary_color:
          type: string
          nullable: true
          maxLength: 9
        brand_accent_color:
          type: string
          nullable: true
          maxLength: 9
        remove_branding_banner:
          type: boolean
          writeOnly: true
          default: false
      required:
      - created_at
      - created_by
      - id
      - name
      - organization
      - updated_at
...
    ShareLink:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        document:
          type: string
          nullable: true
        dataroom:
          type: string
          nullable: true
        document_name:
          type: string
          readOnly: true
          nullable: true
        document_type:
          type: string
          readOnly: true
          nullable: true
        dataroom_name:
          type: string
          readOnly: true
          nullable: true
        dataroom_settings:
          type: array
          items:
            $ref: '#/components/schemas/ShareLinkDataroomSetting'
          readOnly: true
        created_by:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          readOnly: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        has_password:
          type: boolean
          description: Returns True if the link is password-protected.
          readOnly: true
        password:
          type: string
          nullable: true
          maxLength: 255
        requires_email:
          type: boolean
        requires_email_verification:
          type: boolean
        allow_download:
          type: boolean
        enable_watermark:
          type: boolean
        watermark_text:
          type: string
          maxLength: 255
        receive_email_notification:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        view_count:
          type: integer
          description: Returns the number of view sessions for the link.
          readOnly: true
        recent_view_sessions:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: Returns up to 10 most recent view sessions.
          readOnly: true
        last_viewed_at:
          type: string
          format: date-time
          nullable: true
          description: Returns the timestamp of the most recent view session.
          readOnly: true
      required:
      - created_at
      - created_by
      - dataroom_name
      - dataroom_settings
      - document_name
      - document_type
      - has_password
      - id
      - last_viewed_at
      - name
      - recent_view_sessions
      - slug
      - updated_at
      - view_count
    ShareLinkDataroomSetting:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        dataroom_document:
          type: string
          nullable: true
        dataroom_folder:
          type: string
          nullable: true
        is_visible:
          type: boolean
        allow_download:
          type: boolean
        enable_watermark:
          type: boolean
      required:
      - id
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT